Refactor main.rs; Update README; Fix app close; Fix pyrefly issues
